+++ examples/05-scale-showcase/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @cotal-ai/example-05-scale-showcase
 
+## 0.0.14
+
+### Patch Changes
+
+- @cotal-ai/core@0.33.8
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [576ac7d]
+  - @cotal-ai/core@0.33.7
+
 ## 0.0.12
 
 ### Patch Changes