+++ extensions/connector-core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @cotal-ai/connector-core
 
+## 0.33.8
+
+## 0.33.7
+
 ## 0.33.6
 
 ### Patch Changes