+++ implementations/delivery/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @cotal-ai/delivery
 
+## 0.33.8
+
+### Patch Changes
+
+- @cotal-ai/core@0.33.8
+- @cotal-ai/workspace@0.33.8
+
+## 0.33.7
+
+### Patch Changes
+
+- Updated dependencies [576ac7d]
+- Updated dependencies [7119b4c]
+  - @cotal-ai/core@0.33.7
+  - @cotal-ai/workspace@0.33.7
+
 ## 0.33.6
 
 ### Patch Changes