+++ implementations/runtime/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @cotal-ai/runtime
 
+## 0.33.8
+
+### Patch Changes
+
+- @cotal-ai/core@0.33.8
+- @cotal-ai/lang@0.33.8
+
+## 0.33.7
+
+### Patch Changes
+
+- Updated dependencies [576ac7d]
+  - @cotal-ai/core@0.33.7
+  - @cotal-ai/lang@0.33.7
+
 ## 0.33.6
 
 ### Patch Changes